Added public holidays
Added public holidays so that it won't count the business hours that land on or pass through any public holiday. I added the public holidays of Turkey because I Iive in Turkey and use this class for here.
The country can be changed easily from is_holiday(), which can be found at the very bottom.